#9318 : Refactoring JWT-related and new utility classes so the java packages point to "dotcms" instead of "dotmarketing".
