+++ src/com/dotcms/filters/interceptor/jwt/JsonWebTokenInterceptor.java
@@ -1,4 +1,4 @@
-package com.dotmarketing.filters.interceptor.jwt;
+package com.dotcms.filters.interceptor.jwt;
 
 import java.io.IOException;
 import java.util.Date;
@@ -7,23 +7,23 @@ import javax.servlet.ServletRequest;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.dotcms.auth.providers.jwt.beans.DotCMSSubjectBean;
+import com.dotcms.auth.providers.jwt.beans.JWTBean;
+import com.dotcms.auth.providers.jwt.factories.JsonWebTokenFactory;
+import com.dotcms.auth.providers.jwt.services.JsonWebTokenService;
+import com.dotcms.cms.login.LoginService;
+import com.dotcms.cms.login.LoginServiceFactory;
+import com.dotcms.filters.interceptor.WebInterceptor;
+import com.dotcms.util.marshal.MarshalFactory;
+import com.dotcms.util.marshal.MarshalUtils;
+import com.dotcms.util.security.Encryptor;
+import com.dotcms.util.security.EncryptorFactory;
 import com.dotmarketing.business.APILocator;
-import com.dotmarketing.cms.login.LoginService;
-import com.dotmarketing.cms.login.LoginServiceFactory;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
-import com.dotmarketing.filters.interceptor.WebInterceptor;
 import com.dotmarketing.util.Config;
 import com.dotmarketing.util.Logger;
 import com.dotmarketing.util.UtilMethods;
-import com.dotmarketing.util.jwt.DotCMSSubjectBean;
-import com.dotmarketing.util.jwt.JWTBean;
-import com.dotmarketing.util.jwt.JsonWebTokenFactory;
-import com.dotmarketing.util.jwt.JsonWebTokenService;
-import com.dotmarketing.util.marshal.MarshalFactory;
-import com.dotmarketing.util.marshal.MarshalUtils;
-import com.dotmarketing.util.security.Encryptor;
-import com.dotmarketing.util.security.EncryptorFactory;
 import com.liferay.portal.PortalException;
 import com.liferay.portal.SystemException;
 import com.liferay.portal.ejb.CompanyLocalManager;
