+++ src/com/dotcms/util/CollectionsUtils.java
@@ -1,9 +1,16 @@
-package com.dotmarketing.util;
-
-import com.dotcms.repackage.edu.emory.mathcs.backport.java.util.Arrays;
+package com.dotcms.util;
 
 import java.io.Serializable;
-import java.util.*;
+import java.util.AbstractMap;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import com.dotcms.repackage.edu.emory.mathcs.backport.java.util.Arrays;
 
 /**
  * Encapsulates collection utils methods
