+++ src/com/dotcms/util/security/Encryptor.java
@@ -1,15 +1,11 @@
-package com.dotmarketing.util.security;
-
-import com.dotmarketing.cms.factories.PublicCompanyFactory;
-import com.dotmarketing.cms.factories.PublicEncryptionFactory;
-import com.dotmarketing.exception.DotRuntimeException;
-import com.dotmarketing.util.Logger;
-import com.liferay.portal.model.Company;
-import com.liferay.util.EncryptorException;
+package com.dotcms.util.security;
 
 import java.io.Serializable;
 import java.security.Key;
 import java.security.Provider;
+
+import com.dotmarketing.cms.factories.PublicEncryptionFactory;
+import com.liferay.util.EncryptorException;
 
 /**
  * This class is just a wrapper to encapsulate the
