+++ test/com/dotmarketing/filters/interceptor/jwt/JsonWebTokenInterceptorTest.java
@@ -21,17 +21,18 @@ import org.junit.Test;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 
+import com.dotcms.auth.providers.jwt.beans.DotCMSSubjectBean;
+import com.dotcms.auth.providers.jwt.beans.JWTBean;
+import com.dotcms.auth.providers.jwt.factories.JsonWebTokenFactory;
+import com.dotcms.auth.providers.jwt.services.JsonWebTokenService;
+import com.dotcms.cms.login.LoginService;
+import com.dotcms.filters.interceptor.jwt.JsonWebTokenInterceptor;
+import com.dotcms.util.marshal.MarshalFactory;
+import com.dotcms.util.marshal.MarshalUtils;
+import com.dotcms.util.security.Encryptor;
 import com.dotmarketing.business.UserAPI;
-import com.dotmarketing.cms.login.LoginService;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
-import com.dotmarketing.util.jwt.DotCMSSubjectBean;
-import com.dotmarketing.util.jwt.JWTBean;
-import com.dotmarketing.util.jwt.JsonWebTokenFactory;
-import com.dotmarketing.util.jwt.JsonWebTokenService;
-import com.dotmarketing.util.marshal.MarshalFactory;
-import com.dotmarketing.util.marshal.MarshalUtils;
-import com.dotmarketing.util.security.Encryptor;
 import com.liferay.portal.PortalException;
 import com.liferay.portal.SystemException;
 import com.liferay.portal.ejb.CompanyLocalManager;
